requirements.txt
python-telegram-bot==21.6
httpx>=0.27
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
pyarrow>=15.0
openai>=1.40.0
tiktoken>=0.7.0
pydantic>=2.8
